269 Moves components to their own files
