+++ packages/orchestrator-ui-components/src/components/WfoSummary/WfoSummaryCardList.tsx
@@ -1,28 +1,22 @@
-import React, { FC, ReactNode, useState } from 'react';
+import React, { FC } from 'react';
 
-import Link from 'next/link';
 import { useRouter } from 'next/router';
-import type { UrlObject } from 'url';
 
 import {
     EuiButton,
-    EuiFlexGroup,
     EuiFlexItem,
     EuiHorizontalRule,
-    EuiIcon,
     EuiPanel,
     EuiSpacer,
-    EuiTextColor,
     useEuiScrollBar,
 } from '@elastic/eui';
 
 import { useOrchestratorTheme } from '@/hooks';
 
-export type SummaryCardListItem = {
-    title: string;
-    value: ReactNode;
-    url?: string;
-};
+import {
+    SummaryCardListItem,
+    WfoSummaryCardListItem,
+} from './WfoSummaryCardListItem';
 
 export type SummaryCardButtonConfig = {
     name: string;
@@ -70,7 +64,7 @@ export const WfoSummaryCardList: FC<WfoSummaryCardListProps> = ({
                     >
                         {items?.map((item, index) => (
                             <div key={index}>
-                                <WfoSummaryListItem
+                                <WfoSummaryCardListItem
                                     title={item.title}
                                     value={item.value}
                                     url={item.url}
@@ -95,66 +89,5 @@ export const WfoSummaryCardList: FC<WfoSummaryCardListProps> = ({
                 )}
             </EuiPanel>
         </EuiFlexItem>
-    );
-};
-
-export type WfoSummaryListItemProps = {
-    title: string;
-    value: ReactNode;
-    url?: string;
-};
-
-export type WfoOptionalLinkProps = {
-    children: ReactNode;
-    href?: UrlObject | string;
-};
-
-export const WfoOptionalLink: FC<WfoOptionalLinkProps> = ({
-    children,
-    href,
-}) => {
-    if (!href) {
-        return <span>{children}</span>;
-    }
-
-    return <Link href={href}>{children}</Link>;
-};
-
-export const WfoSummaryListItem: FC<WfoSummaryListItemProps> = ({
-    title,
-    value,
-    url,
-}) => {
-    const [hoverState, setHoverState] = useState(false);
-
-    return (
-        <WfoOptionalLink href={url}>
-            <EuiFlexGroup
-                style={{ paddingBlock: 10 }}
-                onMouseOver={() => setHoverState(true)}
-                onMouseLeave={() => setHoverState(false)}
-                gutterSize="none"
-            >
-                <EuiFlexItem>
-                    <EuiTextColor
-                        color={url ? '#397dc2' : 'black'}
-                        style={{ fontWeight: 500 }}
-                    >
-                        {title}
-                    </EuiTextColor>
-                    <EuiTextColor style={{ fontWeight: 400 }}>
-                        {value}
-                    </EuiTextColor>
-                </EuiFlexItem>
-                <EuiFlexItem
-                    grow={false}
-                    css={{
-                        visibility: hoverState && url ? 'visible' : 'hidden',
-                    }}
-                >
-                    <EuiIcon type="sortRight" color="primary" />
-                </EuiFlexItem>
-            </EuiFlexGroup>
-        </WfoOptionalLink>
     );
 };
